Canonical dropped the JSON cloud image catalog, so use the CSV one.
This means we lose some of the metadata, such as SHA-512 hashes.

This also means we don't have nice programmatic access to the
"-disk1.img" urls, especially in combination of knowing what the
serial number for an image is. Thus, even the server vms need to
switch from the "disk1" partitioned layout to having a filesystem on
the disk directly, with no partition table.